Improve runner
The existing runner implementation is based on goexec which maintains an
unbounded buffer to log the process outputs. While this is useful for
testing, long running process may run into memory issues.

This changes the runner to close the buffer when the processes has been
started correctly and continue to piple logging to stdout

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>