Invalidate cache on unique-key violation
When two nodes try to set the state at the same time, one gets a duplicate error and doesn't update the vault. Then we don't fail, because we know that another process has updated it, but the txidstore cache remains with the same previous value. Now we invalidate the cache and fetch the new value.

Signed-off-by: Alexandros Filios <alexandros.filios@ibm.com>