Retry after deadlock
When multiple replicas are trying to commit the same TX and rewrite the same keys, we get a deadlock error from the DB. Now we retry every time we get a deadlock error until the operation is successful. This means that either the other writes have failed so we can write the new values, or the other writes were successful so we update the state (using the same values).

Signed-off-by: Alexandros Filios <alexandros.filios@ibm.com>