Fix flaky session test
The TestSessionLifecycleConcurrent tests if messages can be enqueued on
a closed session. As `enqueue` should drop messages when a session is
closed, the test checked that the underlying chan is empty. However,
this test is flaky as it depends on the consumer goroutine to consume
all messages.

To simplify this test, `enqueue` now returns true if a message was
enqueued in the session, or false if the message is dropped. The test
can now check the returned value.

Signed-off-by: Marcus Brandenburger <bur@zurich.ibm.com>